backend/requirements.txt
fastapi
uvicorn
sqlmodel
psycopg2-binary
alembic
pydantic-settings
python-dotenv
pytest
